Changed overlays to position relative to the ui
Changed overlays that have been set to dynamic to position relative to screen edges rather than just using a fixed point.
This means that if overlays are positioned next to UI elements (backpack, chat box, etc) they will stay next to them when the client is resized or the config panel is opened/closed.
I implemented this mainly so that the screen markers will stay correctly positioned when the client is resized, although it will apply to other dynamic overlays.